pyproject.toml
[project]
name = "quantum-email-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.128.5",
    "httpx>=0.28.1",
    "kyber-py>=1.2.0",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
